Test.Hspec.Monadic

Description

Deprecated: use Test.Hspec, Test.Hspec.Runner or Test.Hspec.Core instead

Types

type Spec = SpecM ()

class Example a

A type class for examples.

Defining a spec

describe :: String -> Spec -> Spec

Combine a list of specs into a larger spec.

context :: String -> Spec -> Spec

An alias for describe.

it :: Example a => String -> a -> Spec

Create a spec item.

A spec item consists of:

a textual description of a desired behavior

an example for that behavior

 describe "absolute" $ do
   it "returns a positive number when given a negative number" $
     absolute (-1) == 1

pending :: Expectation

Specifies a pending example.

If you want to textually specify a behavior but do not have an example yet, use this:

 describe "fancyFormatter" $ do
   it "can format text in a way that everyone likes" $
     pending

Running a spec

hspec :: Spec -> IO ()

Run given spec and write a report to stdout. Exit with exitFailure if at least one spec item fails.

data Summary

Summary of a test run.

Constructors

Summary

Fields

summaryExamples :: Int

summaryFailures :: Int

Interface to the non-monadic API

runSpecM :: Spec -> [SpecTree]

Convert a Spec to a forest of SpecTrees.

fromSpecList :: [SpecTree] -> Spec

Create a Spec from a forest of SpecTrees.

Deprecated types and functions

type Specs = Spec

descriptions :: [Spec] -> Spec

hspecB :: Spec -> IO Bool

hspecX :: Spec -> IO ()

hHspec :: Handle -> Spec -> IO Summary