Test.Hspec.QuickCheck

Description

Params

modifyMaxSuccess :: (Int -> Int) -> Spec -> Spec

Use a modified maxSuccess for given spec.

modifyMaxDiscardRatio :: (Int -> Int) -> Spec -> Spec

Use a modified maxDiscardRatio for given spec.

modifyMaxSize :: (Int -> Int) -> Spec -> Spec

Use a modified maxSize for given spec.

Re-exports from QuickCheck

Previous versions of Hspec provided a distinct property combinator, but it's now possible to use QuickCheck's property instead. For backward compatibility we now re-export QuickCheck's property, but it is advisable to import it from Test.QuickCheck instead.

property :: Testable prop => prop -> Property

Shortcuts

prop :: Testable prop => String -> prop -> Spec

 prop ".." $
   ..

is a shortcut for

 it ".." $ property $
   ..