Test.Hspec.Runner

Description

Running a spec

hspec :: Spec -> IO ()

Run given spec and write a report to stdout. Exit with exitFailure if at least one spec item fails.

hspecResult :: Spec -> IO Summary

Run given spec and returns a summary of the test run.

Note: hspecResult does not exit with exitFailure on failing spec items. If you need this, you have to check the Summary yourself and act accordingly.

hspecWith :: Config -> Spec -> IO Summary

Run given spec with custom options and returns a summary of the test run.

Note: hspecWith does not exit with exitFailure on failing spec items. If you need this, you have to check the Summary yourself and act accordingly.

Types

data Summary

Summary of a test run.

Constructors

Summary

Fields

summaryExamples :: Int

summaryFailures :: Int

data Config

Constructors

Config

Fields

configDryRun :: Bool

configPrintCpuTime :: Bool

configFastFail :: Bool

configFilterPredicate :: Maybe (Path -> Bool)

A predicate that is used to filter the spec before it is run. Only examples that satisfy the predicate are run.

configQuickCheckSeed :: Maybe Integer

configQuickCheckMaxSuccess :: Maybe Int

configQuickCheckMaxDiscardRatio :: Maybe Int

configQuickCheckMaxSize :: Maybe Int

configSmallCheckDepth :: Int

configColorMode :: ColorMode

configFormatter :: Formatter

configHtmlOutput :: Bool

configHandle :: Either Handle FilePath

data ColorMode

Constructors

ColorAuto

ColorNever

ColorAlways

type Path = ([String], String)

A tuple that represents the location of an example within a spec.

It consists of a list of group descriptions and a requirement description.

defaultConfig :: Config

configAddFilter :: (Path -> Bool) -> Config -> Config

Add a filter predicate to config. If there is already a filter predicate, then combine them with ||.

Internals

hspecWithFormatter :: IsFormatter a => a -> Spec -> IO ()

This function is used by hspec-discover. It is not part of the public API and may change at any time.