Test.Hspec.Core

Description

This module provides access to Hspec's internals. It is less stable than other parts of the API. For most users Test.Hspec is more suitable!

A type class for examples

class Example a where

A type class for examples.

Methods

evaluateExample :: a -> Params -> (IO () -> IO ()) -> IO Result

data Params

Constructors

Params

Fields

paramsQuickCheckArgs :: Args

paramsSmallCheckDepth :: Int

paramsReportProgress :: Progress -> IO ()

type Progress = (Int, Int)

data Result

The result of running an example.

Constructors

Success

Pending (Maybe String)

Fail String

A writer monad for constructing specs

data SpecM a

A writer monad for SpecTree forests.

runSpecM :: Spec -> [SpecTree]

Convert a Spec to a forest of SpecTrees.

fromSpecList :: [SpecTree] -> Spec

Create a Spec from a forest of SpecTrees.

Internal representation of a spec tree

data SpecTree

Internal representation of a spec.

Constructors

SpecGroup String [SpecTree]

SpecItem Item

data Item

Constructors

Item

Fields

itemIsParallelizable :: Bool

itemRequirement :: String

itemExample :: Params -> (IO () -> IO ()) -> IO Result

mapSpecItem :: (Item -> Item) -> Spec -> Spec

modifyParams :: (Params -> Params) -> Spec -> Spec

describe :: String -> [SpecTree] -> SpecTree

The describe function combines a list of specs into a larger spec.

it :: Example a => String -> a -> SpecTree

Create a spec item.

Deprecated types and functions

type Specs = [SpecTree]

Deprecated: use `[SpecTree]` instead

hspecB :: [SpecTree] -> IO Bool

Deprecated: use hspecWith instead

hspecX :: [SpecTree] -> IO ()

Deprecated: use hspec instead

hHspec :: Handle -> [SpecTree] -> IO Summary

Deprecated: use hspecWith instead

hspec :: [SpecTree] -> IO ()

Deprecated: use hspec instead