Test.Hspec.Discover

Description

Warning: This module is used by hspec-discover. It is not part of the public API and may change at any time.

Documentation

type Spec = SpecWith ()

hspec :: Spec -> IO ()

Run a given spec and write a report to stdout. Exit with exitFailure if at least one spec item fails.

Note: hspec handles command-line options and reads config files. This is not always desired. Use runSpec if you need more control over these aspects.

class IsFormatter a where

Minimal complete definition

toFormatter

Methods

toFormatter :: a -> IO Formatter

hspecWithFormatter :: IsFormatter a => a -> Spec -> IO ()

postProcessSpec :: FilePath -> Spec -> Spec

describe :: HasCallStack => String -> SpecWith a -> SpecWith a

The describe function combines a list of specs into a larger spec.

module Prelude